Moved user search API and test cases to routes/ShortcutSearch.js and test/api/ShortcutSearch.js
